Moved code to base class
