+++ openghg/modules/_obs_surface.py
@@ -12,28 +12,9 @@ class ObsSurface(BaseModule):
     _root = "ObsSurface"
     _uuid = "da0b8b44-6f85-4d3c-b6a3-3dde34f6dea1"
 
-    def __init__(self):
-        super().__init__()
-
-    def to_data(self) -> Dict:
-        """Return a JSON-serialisable dictionary of object
-        for storage in object store
-
-        Returns:
-            dict: Dictionary version of object
-        """
-        from Acquire.ObjectStore import datetime_to_string
-
-        data = {}
-        data["creation_datetime"] = datetime_to_string(self._creation_datetime)
-        data["stored"] = self._stored
-        data["datasource_table"] = self._datasource_table
-        data["datasource_uuids"] = self._datasource_uuids
-        data["datasource_names"] = self._datasource_names
-        data["file_hashes"] = self._file_hashes
-        data["rank_data"] = self._rank_data
-
-        return data
+    # We don't currently need to add anything here
+    # def __init__(self):
+    #     super().__init__()
 
     def save(self, bucket: Optional[Dict] = None) -> None:
         """Save the object to the object store
@@ -231,8 +212,4 @@ class ObsSurface(BaseModule):
         key = f"{Datasource._datasource_root}/uuid/{uuid}"
         delete_object(bucket=bucket, key=key)
 
-        # First remove from our dictionary of Datasources
-        name = self._datasource_uuids[uuid]
-
-        del self._datasource_names[name]
         del self._datasource_uuids[uuid]
